crm_for_students/requirements/dev.txt
# Core Django Framework
Django==5.0
asgiref==3.10.0
sqlparse==0.5.3

# Database
psycopg2-binary==2.9.11

# Admin UI
django-unfold

# CORS
django-cors-headers

# Environment & Configuration
python-decouple==3.8
python-dotenv==1.1.1

# Build Tools
setuptools==80.9.0

